requirements.txt
Django==2.0.5
psycopg2==2.7.4
psycopg2-binary
pytz==2018.4
djangorestframework
markdown
django-filter
Celery==4.1.0
amqp==2.2.2
django-redis
msgpack-python
